creates a centerline from the raster based corridor mask

The centerline function derives the centerline of the river corridor mask

Usage

centerline(mask, riverline, seg_length = 1000, smooth_factor = 2500, d = 5000)

Arguments

mask

spatRaster of the binary river corridor mask where the river corridor is encoded with 1 and non-corridor is NA; alternatively, mask can also be a sf object with a polygon representing the river corridor

riverline

sf object representing the river line

seg_length

length of the desired river segments; choosing the appropriate scale of analysis will depend very much on the resolution of the input data and the scope of the specific analysis; default is 1 km

smooth_factor

smoothing factor for converting the raster mask to polygon, a higher smooth parameter will result in higher degree of simplification of the polygon

d

length to extend the ends of the riverline in order to split the corridor outline; can be adjusted to avoid errors

Value

a list with 1) the centerline as sf line object and 2) the smoothed polygon as sf object