Compute Membership based on Linear Functions

Function to create linear fuzzy membership functions to delineate river corridor. The function can be either linear or inverse linear (see details).

Usage

computeMembership(index, x_cal_min, x_cal_max, type = "linear")

Arguments

index

spat raster or path to raster file in any format readable by terra::rast()

x_cal_min

numeric; lower bound of calibration range

x_cal_max

numeric; upper bound of calibration range

type

character; can be either "linear" or "inv_linear"

Value

spatRaster containing the membership raster for the specific index

References

Betz, F., Lauermann, M., Cyffka, B. (2018): Delineation of the riparian zone in data-scarce regions using fuzzy membership functions: An evaluation based on the case of the Naryn River in Kyrgyzstan. Geomorphology. https://doi.org/10.1016/j.geomorph.2018.01.024

Author

Florian Betz