Extracting a channel network from a digital elevation network using GRASS GIS

This function provides a wrapper for two different approaches for channel network delineation from a digital elevation model

Usage

extract_channels_grass(
  dem,
  preprocess_dem = FALSE,
  grass_path,
  grass_db,
  basin_threshold,
  channel_threshold = 10000,
  montgomery = 0,
  mfd_threshold = 0,
  min_seglength,
  compute_orders = TRUE,
  stream_orders = c("hack", "strahler"),
  disk_swap = FALSE,
  memory = 300,
  remove = TRUE
)

Arguments

dem

Path to a digital elevation model stored in a common raster format or terra SpatRaster

preprocess_dem

should the DEM be hydrologically conditioned before analysis? If TRUE, a minimum impact DEM treatment using the procedure of Lindsay and Creed (2005) is carried out.

grass_path

path where the GRASS installation is found (see rgrass::initGRASS for details)

grass_db

directory where to create the GRASS database; if not provided, tempdir() is used instead

basin_threshold

integer; minimum size of exterior watershed basin (see the help of r.watershed for more details)

channel_threshold

flow accumulation threshold for initiating channels; if montgomery is>0, the channel initiation threshold method according to Montgomery and Foufoula-Georgiou (1993) is used instead of flow accumulation

montgomery

exponent for adjusting flow accumulation to slope

mfd_threshold

flow accumulation threshold where to switch from multiple flow direction to single flow direction for channel derivation; setting mfd_threshold to 0 (the default) disables multiple flow direction completely

min_seglength

length (in cells) which a channel segment needs to have at least to be returned

compute_orders

logical; if TRUE, commonly used stream orders will be computed using the r.stream.order addon of GRASS GIS

stream_orders

character vector of the stream orders to return in the stream raster output; can be any strahler, horton, hack, shreve, topo; by GRASS GIS default, the stream vector output will automatically include all orders as attributes

disk_swap

logical; if FALSE, then the computation will be carried out all in RAM; if TRUE, disk swap calculation is used consuming a maximum amount of RAM is specified in memory

memory

maximum RAM to use for disk swap computation, default is 300

remove

logical; if TRUE, the temporary GRASS database will be removed when quitting the R session or unloading rgrass

Value

a list with a raster (terra rast() format) and a vector (sf() format) dataset of the channel network

References

Lindsay, J. B., and Creed, I. F. 2005. Removal of artifact depressions from digital elevation models: towards a minimum impact approach. Hydrological Processes 19, 3113-3126. DOI: 10.1002/hyp.5835

Author

Florian Betz