Computing distance along a channel

This function computes the distance along a channel line, where the line is "exploded" before computing distances. The function assumes the channel line being in projected coordinates and will give an error if this is not the case.

Usage

profile_distance(x, maxDist = 10)

Arguments

x

sf object containing the channel line

maxDist

maximum distance (in meter) between two consecutive vertices when densifying the vertices of the line. Default is 10 m.In general, maxDist should be in a similar range as the resolution of the underlying elevation data.

Author

Florian Betz